+++ test/form/samples/no-treeshake/_config.js
@@ -5,7 +5,8 @@ module.exports = defineRollupTest({
 		treeshake: false,
 		output: {
 			globals: { external: 'external' },
-			name: /* not shaken, but */ 'stirred'
+			name: /* not shaken, but */ 'stirred',
+			inlineDynamicImports: true
 		}
 	}
 });