+++ test/form/samples/treeshake-deterministic-dynamic-import/sub4.js
@@ -1,0 +1,4 @@
+export const foo4 = 3;
+export const bar4 = 4;
+
+export * from 'external';